Consolidate plugin cache into a single module
Move plugin cache restore and save into src/plugin-cache.js, which now
owns the enable check, plugin-dir resolution, cache-key derivation, and
the previously-shared state keys (now private). Extract resolvePluginDir
and cacheKey as pure, testable helpers.

cache-restore.js and cache-save.js become thin adapters re-exporting
restore/save so setup-tflint.js and post.js are untouched.
